Fundamentalt omarbetat hur tid fungerar och rört typ allt.
OBS! Använder inte Håkans exempel vid push av testningsanledningar.
Fungerar dock inte enligt exemplet ännu.

Tid betraktas nu som att när ett event inträffar är den globala tiden
den tiden, och ankomsttider/plocktider/etc betraktas som ett tidsdelta.
Det vill säga, om den första kunden anländer 0.44 så är dess händelsetid
globala tiden (0) plus kundens ankomsttid (0.44). Varje händelse gör
därmed en tilldelning till den globala tiden. Detta är gemensamt för
alla events.

AnkomstEvent.java: Minimala ändringar i konstruktorn. Execute-metoden
omarbetad till att fungera på ett något annorlunda sätt. Nytt
ankomstevent genereras nu i slutet och bara om butiken är öppen. La till
en hjälpmetod canEnter.

BetalaEvent.java: queueTimer spåras nu i kundens objekt. HarKassa
används inte längre; har ersatts av kontroller på ledigakassor.
timeChange har tagits bort på grund av omarbetningen av tid. Har nu två
nya hjälpmetoder. queueTime() lägger till kundens betalningstid som
väntetid på alla väntande kunder. nextInQueue kontrollerar om det finns
kunder i kön. Om det finns kunder i kön skapas ett nytt betalaevent med
den kunden och kunden tas bort från kön, annars returneras false vilket
ökar mängden lediga kassor.

Event.java: timeChange har tagits bort.

EventQueue.java: Har fått en debug-rad.

Kund.java: Main-metoden togs bort. queueTimer initialiseras nu i
konstruktorn.

MarketEvent.java: harKassa togs bort då den inte längre används.
runNextEvent() har reducerats kraftigt då en del kod var specifik för
ett event, vilket i så fall borde överskuggas i det eventet. Metoden kör
nu bara nästa event om kön inte är tom och run är true. Metoden
newAnkomst togs bort då den är redundant.

MarketView.java: Använder nu eventets händelsetid i sin utskrift
istället för MarketStates. Visar nu alltid två decimaler.

PlockEvent.java: Kontrollerar nu via en hjälpmetod om kassorna är
upptagna. Om alla kassor är upptagna lägger hjälpmetoden till kunden i
kassakön och ökar räknaren för unika köande kunder. Annars skapas ett
nytt betalaevent och antalet lediga kassor minskas.

StartEvent.java: Tog bort main-metoden. Tar nu bort sig själv från kön i
sin execute.

State.java: Tar nu en double som parameter i sin konstruktor.

StopEvent.java: Tar nu bort sig själv från kön.

+++ lab5/src/lab5/classtemplates/state/State.java
@@ -19,7 +19,7 @@ public class State extends Observable {
 	 * 
 	 * @param time Starttiden för tillståndet.
 	 */
-	public State(int time) {
+	public State(double time) {
 		timeMax = time;
 		run = true;
 	}
